Convert Logo to separate file
